+++ profiling/profile_prev_succ.py
@@ -11,7 +11,7 @@ from timeit import Timer
 from random import shuffle
 
 from bintrees import RBTree as PTree
-from bintrees import FastRBTree as FTree
+from bintrees import FastRBTree as FTree, has_fast_tree_support
 
 COUNT = 100
 
@@ -93,4 +93,8 @@ def main():
 
 
 if __name__ == '__main__':
+    if not has_fast_tree_support():
+        print("Cython extension for FastRBTree is NOT working.")
+    else:
+        print("Cython extension for FastRBTree is working.")
     main()